+++ lib/AmdTemplatePlugin.js
@@ -7,9 +7,9 @@
 
 const { ConcatSource } = require("webpack-sources");
 const ExternalModule = require("./ExternalModule");
-const JavascriptModulesPlugin = require("./JavascriptModulesPlugin");
 const RuntimeGlobals = require("./RuntimeGlobals");
 const Template = require("./Template");
+const JavascriptModulesPlugin = require("./javascript/JavascriptModulesPlugin");
 
 /** @typedef {import("./Compiler")} Compiler */
