+++ lib/ChunkTemplate.js
@@ -11,7 +11,7 @@ const memorize = require("./util/memorize");
 /** @typedef {import("./Compilation")} Compilation */
 
 const getJavascriptModulesPlugin = memorize(() =>
-	require("./JavascriptModulesPlugin")
+	require("./javascript/JavascriptModulesPlugin")
 );
 
 // TODO webpack 6 remove this class
