+++ lib/CompatibilityPlugin.js
@@ -8,7 +8,7 @@
 const ConstDependency = require("./dependencies/ConstDependency");
 
 /** @typedef {import("./Compiler")} Compiler */
-/** @typedef {import("./JavascriptParser")} JavascriptParser */
+/** @typedef {import("./javascript/JavascriptParser")} JavascriptParser */
 
 const nestedWebpackRequireTag = Symbol("nested __webpack_require__");
