+++ lib/DefinePlugin.js
@@ -15,7 +15,7 @@ const RuntimeGlobals = require("./RuntimeGlobals");
 const ConstDependency = require("./dependencies/ConstDependency");
 
 /** @typedef {import("./Compiler")} Compiler */
-/** @typedef {import("./JavascriptParser")} JavascriptParser */
+/** @typedef {import("./javascript/JavascriptParser")} JavascriptParser */
 /** @typedef {null|undefined|RegExp|Function|string|number} CodeValuePrimitive */
 /** @typedef {CodeValuePrimitive|Record<string, CodeValuePrimitive>|RuntimeValue} CodeValue */
