+++ lib/EvalDevToolModulePlugin.js
@@ -6,8 +6,8 @@
 "use strict";
 
 const { RawSource } = require("webpack-sources");
-const JavascriptModulesPlugin = require("./JavascriptModulesPlugin");
 const ModuleFilenameHelpers = require("./ModuleFilenameHelpers");
+const JavascriptModulesPlugin = require("./javascript/JavascriptModulesPlugin");
 
 /** @typedef {import("webpack-sources").Source} Source */
 /** @typedef {import("./Compiler")} Compiler */
