+++ lib/EvalSourceMapDevToolPlugin.js
@@ -6,10 +6,10 @@
 "use strict";
 
 const { RawSource } = require("webpack-sources");
-const JavascriptModulesPlugin = require("./JavascriptModulesPlugin");
 const ModuleFilenameHelpers = require("./ModuleFilenameHelpers");
 const NormalModule = require("./NormalModule");
 const SourceMapDevToolModuleOptionsPlugin = require("./SourceMapDevToolModuleOptionsPlugin");
+const JavascriptModulesPlugin = require("./javascript/JavascriptModulesPlugin");
 
 /** @typedef {import("webpack-sources").Source} Source */
 /** @typedef {import("./Compiler")} Compiler */
