+++ lib/ExportPropertyTemplatePlugin.js
@@ -6,8 +6,8 @@
 "use strict";
 
 const { ConcatSource } = require("webpack-sources");
-const JavascriptModulesPlugin = require("./JavascriptModulesPlugin");
 const { UsageState } = require("./ModuleGraph");
+const JavascriptModulesPlugin = require("./javascript/JavascriptModulesPlugin");
 
 /** @typedef {import("./Compiler")} Compiler */
