+++ lib/ExportsInfoApiPlugin.js
@@ -9,7 +9,7 @@ const ConstDependency = require("./dependencies/ConstDependency");
 const ExportsInfoDependency = require("./dependencies/ExportsInfoDependency");
 
 /** @typedef {import("./Compiler")} Compiler */
-/** @typedef {import("./JavascriptParser")} JavascriptParser */
+/** @typedef {import("./javascript/JavascriptParser")} JavascriptParser */
 
 class ExportsInfoApiPlugin {
 	/**
