+++ lib/HotModuleReplacementPlugin.js
@@ -8,7 +8,6 @@
 const { SyncBailHook } = require("tapable");
 const { RawSource } = require("webpack-sources");
 const HotUpdateChunk = require("./HotUpdateChunk");
-const JavascriptParser = require("./JavascriptParser");
 const {
 	evaluateToIdentifier,
 	evaluateToString,
@@ -20,6 +19,7 @@ const ModuleHotAcceptDependency = require("./dependencies/ModuleHotAcceptDepende
 const ModuleHotDeclineDependency = require("./dependencies/ModuleHotDeclineDependency");
 const ModuleHotDependency = require("./dependencies/ModuleHotDependency");
 const HotModuleReplacementRuntimeModule = require("./hmr/HotModuleReplacementRuntimeModule");
+const JavascriptParser = require("./javascript/JavascriptParser");
 const { find } = require("./util/SetHelpers");
 const { compareModulesById } = require("./util/comparators");
