+++ lib/MainTemplate.js
@@ -27,7 +27,7 @@ const memorize = require("./util/memorize");
 /** @typedef {import("./Template").RenderManifestEntry} RenderManifestEntry} */
 
 const getJavascriptModulesPlugin = memorize(() =>
-	require("./JavascriptModulesPlugin")
+	require("./javascript/JavascriptModulesPlugin")
 );
 
 // TODO webpack 6 remove this class
