+++ lib/ModuleInfoHeaderPlugin.js
@@ -6,8 +6,8 @@
 "use strict";
 
 const { ConcatSource } = require("webpack-sources");
-const JavascriptModulesPlugin = require("./JavascriptModulesPlugin");
 const Template = require("./Template");
+const JavascriptModulesPlugin = require("./javascript/JavascriptModulesPlugin");
 
 /** @typedef {import("./Compiler")} Compiler */
 /** @typedef {import("./ModuleTemplate")} ModuleTemplate */
