+++ lib/ModuleTemplate.js
@@ -19,7 +19,7 @@ const memorize = require("./util/memorize");
 /** @typedef {import("./util/Hash")} Hash */
 
 const getJavascriptModulesPlugin = memorize(() =>
-	require("./JavascriptModulesPlugin")
+	require("./javascript/JavascriptModulesPlugin")
 );
 
 /**
