+++ lib/SetVarTemplatePlugin.js
@@ -6,8 +6,8 @@
 "use strict";
 
 const { ConcatSource } = require("webpack-sources");
-const JavascriptModulesPlugin = require("./JavascriptModulesPlugin");
 const RuntimeGlobals = require("./RuntimeGlobals");
+const JavascriptModulesPlugin = require("./javascript/JavascriptModulesPlugin");
 
 /** @typedef {import("./Compiler")} Compiler */
