+++ lib/UmdTemplatePlugin.js
@@ -7,9 +7,9 @@
 
 const { ConcatSource, OriginalSource } = require("webpack-sources");
 const ExternalModule = require("./ExternalModule");
-const JavascriptModulesPlugin = require("./JavascriptModulesPlugin");
 const RuntimeGlobals = require("./RuntimeGlobals");
 const Template = require("./Template");
+const JavascriptModulesPlugin = require("./javascript/JavascriptModulesPlugin");
 
 /** @typedef {import("webpack-sources").Source} Source */
 /** @typedef {import("../declarations/WebpackOptions").LibraryCustomUmdCommentObject} LibraryCustomUmdCommentObject */
