+++ lib/WebpackOptionsApply.js
@@ -7,8 +7,8 @@
 
 const OptionsApply = require("./OptionsApply");
 
-const JavascriptModulesPlugin = require("./JavascriptModulesPlugin");
-const JsonModulesPlugin = require("./JsonModulesPlugin");
+const JavascriptModulesPlugin = require("./javascript/JavascriptModulesPlugin");
+const JsonModulesPlugin = require("./json/JsonModulesPlugin");
 
 const LoaderTargetPlugin = require("./LoaderTargetPlugin");
