+++ lib/debug/ProfilingPlugin.js
@@ -308,7 +308,7 @@ const interceptAllParserHooks = (moduleFactory, tracer) => {
 const interceptAllJavascriptModulesPluginHooks = (compilation, tracer) => {
 	interceptAllHooksFor(
 		{
-			hooks: require("../JavascriptModulesPlugin").getCompilationHooks(
+			hooks: require("../javascript/JavascriptModulesPlugin").getCompilationHooks(
 				compilation
 			)
 		},
