+++ lib/dependencies/ContextDependencyHelpers.js
@@ -6,7 +6,7 @@
 "use strict";
 
 /** @typedef {import("../BasicEvaluatedExpression")} BasicEvaluatedExpression */
-/** @typedef {import("../JavascriptParser")} JavascriptParser */
+/** @typedef {import("../javascript/JavascriptParser")} JavascriptParser */
 /** @typedef {import("./ContextDependency")} ContextDependency */
 /** @typedef {import("./ContextDependency").ContextDependencyOptions} ContextDependencyOptions */
