+++ lib/index.js
@@ -56,7 +56,8 @@ exportPlugins(module.exports, {
 	Generator: () => require("./Generator"),
 	HotModuleReplacementPlugin: () => require("./HotModuleReplacementPlugin"),
 	IgnorePlugin: () => require("./IgnorePlugin"),
-	JavascriptModulesPlugin: () => require("./JavascriptModulesPlugin"),
+	JavascriptModulesPlugin: () =>
+		require("./javascript/JavascriptModulesPlugin"),
 	LibManifestPlugin: () => require("./LibManifestPlugin"),
 	LibraryTemplatePlugin: () => require("./LibraryTemplatePlugin"),
 	LoaderOptionsPlugin: () => require("./LoaderOptionsPlugin"),
