+++ lib/javascript/JavascriptGenerator.js
@@ -7,17 +7,17 @@
 
 const util = require("util");
 const { ConcatSource, RawSource, ReplaceSource } = require("webpack-sources");
-const Generator = require("./Generator");
+const Generator = require("../Generator");
 
 /** @typedef {import("webpack-sources").Source} Source */
-/** @typedef {import("./DependenciesBlock")} DependenciesBlock */
-/** @typedef {import("./Dependency")} Dependency */
-/** @typedef {import("./DependencyTemplates")} DependencyTemplates */
-/** @typedef {import("./Generator").GenerateContext} GenerateContext */
-/** @typedef {import("./InitFragment")} InitFragment */
-/** @typedef {import("./Module")} Module */
-/** @typedef {import("./NormalModule")} NormalModule */
-/** @typedef {import("./RuntimeTemplate")} RuntimeTemplate */
+/** @typedef {import("../DependenciesBlock")} DependenciesBlock */
+/** @typedef {import("../Dependency")} Dependency */
+/** @typedef {import("../DependencyTemplates")} DependencyTemplates */
+/** @typedef {import("../Generator").GenerateContext} GenerateContext */
+/** @typedef {import("../InitFragment")} InitFragment */
+/** @typedef {import("../Module")} Module */
+/** @typedef {import("../NormalModule")} NormalModule */
+/** @typedef {import("../RuntimeTemplate")} RuntimeTemplate */
 
 // TODO: clean up this file
 // replace with newer constructs
