+++ lib/javascript/JavascriptModulesPlugin.js
@@ -12,27 +12,27 @@ const {
 	PrefixSource,
 	RawSource
 } = require("webpack-sources");
-const Compilation = require("./Compilation");
-const { tryRunOrWebpackError } = require("./HookWebpackError");
-const HotUpdateChunk = require("./HotUpdateChunk");
+const Compilation = require("../Compilation");
+const { tryRunOrWebpackError } = require("../HookWebpackError");
+const HotUpdateChunk = require("../HotUpdateChunk");
+const RuntimeGlobals = require("../RuntimeGlobals");
+const Template = require("../Template");
+const { compareModulesByIdentifier } = require("../util/comparators");
+const createHash = require("../util/createHash");
 const JavascriptGenerator = require("./JavascriptGenerator");
 const JavascriptParser = require("./JavascriptParser");
-const RuntimeGlobals = require("./RuntimeGlobals");
-const Template = require("./Template");
-const { compareModulesByIdentifier } = require("./util/comparators");
-const createHash = require("./util/createHash");
 
 /** @typedef {import("webpack-sources").Source} Source */
-/** @typedef {import("./Chunk")} Chunk */
-/** @typedef {import("./ChunkGraph")} ChunkGraph */
-/** @typedef {import("./Compilation").ChunkHashContext} ChunkHashContext */
-/** @typedef {import("./Compiler")} Compiler */
-/** @typedef {import("./DependencyTemplates")} DependencyTemplates */
-/** @typedef {import("./Module")} Module */
-/** @typedef {import("./Module").CodeGenerationResult} CodeGenerationResult */
-/** @typedef {import("./ModuleGraph")} ModuleGraph */
-/** @typedef {import("./RuntimeTemplate")} RuntimeTemplate */
-/** @typedef {import("./util/Hash")} Hash */
+/** @typedef {import("../Chunk")} Chunk */
+/** @typedef {import("../ChunkGraph")} ChunkGraph */
+/** @typedef {import("../Compilation").ChunkHashContext} ChunkHashContext */
+/** @typedef {import("../Compiler")} Compiler */
+/** @typedef {import("../DependencyTemplates")} DependencyTemplates */
+/** @typedef {import("../Module")} Module */
+/** @typedef {import("../Module").CodeGenerationResult} CodeGenerationResult */
+/** @typedef {import("../ModuleGraph")} ModuleGraph */
+/** @typedef {import("../RuntimeTemplate")} RuntimeTemplate */
+/** @typedef {import("../util/Hash")} Hash */
 
 /**
  * @param {Chunk} chunk a chunk
