+++ lib/javascript/JavascriptParser.js
@@ -8,8 +8,8 @@
 const { Parser } = require("acorn");
 const { SyncBailHook, HookMap } = require("tapable");
 const vm = require("vm");
-const BasicEvaluatedExpression = require("./BasicEvaluatedExpression");
-const StackedMap = require("./util/StackedMap");
+const BasicEvaluatedExpression = require("../BasicEvaluatedExpression");
+const StackedMap = require("../util/StackedMap");
 
 /** @typedef {import("estree").ArrayExpression} ArrayExpressionNode */
 /** @typedef {import("estree").BinaryExpression} BinaryExpressionNode */
@@ -38,7 +38,7 @@ const StackedMap = require("./util/StackedMap");
 
 // Syntax: https://developer.mozilla.org/en/SpiderMonkey/Parser_API
 
-const parser = Parser.extend(require("./parsing/importAwaitAcornPlugin"));
+const parser = Parser.extend(require("../parsing/importAwaitAcornPlugin"));
 
 class VariableInfo {
 	/**
