+++ lib/json/JsonGenerator.js
@@ -6,13 +6,13 @@
 "use strict";
 
 const { ConcatSource, RawSource } = require("webpack-sources");
-const Generator = require("./Generator");
-const { UsageState } = require("./ModuleGraph");
-const RuntimeGlobals = require("./RuntimeGlobals");
+const Generator = require("../Generator");
+const { UsageState } = require("../ModuleGraph");
+const RuntimeGlobals = require("../RuntimeGlobals");
 
 /** @typedef {import("webpack-sources").Source} Source */
-/** @typedef {import("./Generator").GenerateContext} GenerateContext */
-/** @typedef {import("./NormalModule")} NormalModule */
+/** @typedef {import("../Generator").GenerateContext} GenerateContext */
+/** @typedef {import("../NormalModule")} NormalModule */
 
 const stringifySafe = data => {
 	const stringified = JSON.stringify(data);
