+++ lib/json/JsonModulesPlugin.js
@@ -8,7 +8,7 @@
 const JsonGenerator = require("./JsonGenerator");
 const JsonParser = require("./JsonParser");
 
-/** @typedef {import("./Compiler")} Compiler */
+/** @typedef {import("../Compiler")} Compiler */
 
 class JsonModulesPlugin {
 	/**
