+++ lib/json/JsonParser.js
@@ -6,9 +6,9 @@
 "use strict";
 
 const parseJson = require("json-parse-better-errors");
-const StaticExportsDependency = require("./dependencies/StaticExportsDependency");
+const StaticExportsDependency = require("../dependencies/StaticExportsDependency");
 
-/** @typedef {import("./NormalModule").ParserState} ParserState */
+/** @typedef {import("../NormalModule").ParserState} ParserState */
 
 class JsonParser {
 	/**
