+++ lib/node/NodeTemplatePlugin.js
@@ -6,9 +6,9 @@
 "use strict";
 
 const { ConcatSource } = require("webpack-sources");
-const JavascriptModulesPlugin = require("../JavascriptModulesPlugin");
 const RuntimeGlobals = require("../RuntimeGlobals");
 const Template = require("../Template");
+const JavascriptModulesPlugin = require("../javascript/JavascriptModulesPlugin");
 const ReadFileChunkLoadingRuntimeModule = require("./ReadFileChunkLoadingRuntimeModule");
 const RequireChunkLoadingRuntimeModule = require("./RequireChunkLoadingRuntimeModule");
