+++ lib/node/ReadFileChunkLoadingRuntimeModule.js
@@ -4,10 +4,10 @@
 
 "use strict";
 
-const chunkHasJs = require("../JavascriptModulesPlugin").chunkHasJs;
 const RuntimeGlobals = require("../RuntimeGlobals");
 const RuntimeModule = require("../RuntimeModule");
 const Template = require("../Template");
+const chunkHasJs = require("../javascript/JavascriptModulesPlugin").chunkHasJs;
 const compileBooleanMatcher = require("../util/compileBooleanMatcher");
 
 class ReadFileChunkLoadingRuntimeModule extends RuntimeModule {
