+++ lib/optimize/ConcatenatedModule.js
@@ -12,7 +12,6 @@ const {
 	ReplaceSource
 } = require("webpack-sources");
 const DependencyTemplate = require("../DependencyTemplate");
-const JavascriptParser = require("../JavascriptParser");
 const Module = require("../Module");
 const { UsageState } = require("../ModuleGraph");
 const RuntimeGlobals = require("../RuntimeGlobals");
@@ -26,6 +25,7 @@ const HarmonyExportSpecifierDependency = require("../dependencies/HarmonyExportS
 const HarmonyImportDependency = require("../dependencies/HarmonyImportDependency");
 const HarmonyImportSideEffectDependency = require("../dependencies/HarmonyImportSideEffectDependency");
 const HarmonyImportSpecifierDependency = require("../dependencies/HarmonyImportSpecifierDependency");
+const JavascriptParser = require("../javascript/JavascriptParser");
 const LazySet = require("../util/LazySet");
 const createHash = require("../util/createHash");
 const contextify = require("../util/identifier").contextify;
