+++ lib/optimize/InnerGraphPlugin.js
@@ -12,7 +12,7 @@ const PureExpressionDependency = require("../dependencies/PureExpressionDependen
 
 /** @typedef {import("../Compiler")} Compiler */
 /** @typedef {import("../Dependency")} Dependency */
-/** @typedef {import("../JavascriptParser")} JavascriptParser */
+/** @typedef {import("../javascript/JavascriptParser")} JavascriptParser */
 
 const topLevelSymbolTag = Symbol("top level symbol");
