+++ lib/web/JsonpTemplatePlugin.js
@@ -9,9 +9,9 @@ const { SyncWaterfallHook } = require("tapable");
 const { ConcatSource } = require("webpack-sources");
 const Compilation = require("../Compilation");
 const HotUpdateChunk = require("../HotUpdateChunk");
-const JavascriptModulesPlugin = require("../JavascriptModulesPlugin");
 const RuntimeGlobals = require("../RuntimeGlobals");
 const Template = require("../Template");
+const JavascriptModulesPlugin = require("../javascript/JavascriptModulesPlugin");
 const JsonpChunkLoadingRuntimeModule = require("./JsonpChunkLoadingRuntimeModule");
 const getEntryInfo = require("./JsonpHelpers").getEntryInfo;
