+++ lib/webworker/WebWorkerTemplatePlugin.js
@@ -7,9 +7,9 @@
 
 const { ConcatSource } = require("webpack-sources");
 const HotUpdateChunk = require("../HotUpdateChunk");
-const JavascriptModulesPlugin = require("../JavascriptModulesPlugin");
 const RuntimeGlobals = require("../RuntimeGlobals");
 const Template = require("../Template");
+const JavascriptModulesPlugin = require("../javascript/JavascriptModulesPlugin");
 const ImportScriptsChunkLoadingRuntimeModule = require("./ImportScriptsChunkLoadingRuntimeModule");
 
 /** @typedef {import("../Compiler")} Compiler */
